requirements.txt
streamlit>=1.24.0
pymongo>=4.4.0
python-dotenv>=1.0.0
matplotlib>=3.7.1
plotly>=5.14.1
pandas>=2.0.2
langchain>=0.0.286
langchain-google-genai>=0.0.1
google-generativeai>=0.3.0
python-decouple>=3.8